area/ui: disable the reset view button when inactive
This commit makes sure that the reset view button isn't active until the icicle graph has been clicked through.
